use models from cyclops/models and delete baseline_models
