--- cyclops/monitor/baseline_models/__init__.py
@@ -1,1 +1,0 @@
-"""Baseline models for static and temporal drift detection for GEMINI use case."""
