--- cyclops/monitor/baseline_models/static/__init__.py
@@ -1,1 +1,0 @@
-"""Static models for drift detection."""
