--- cyclops/monitor/baseline_models/temporal/__init__.py
@@ -1,1 +1,0 @@
-"""Temporal models for drift detection."""
