--- cyclops/monitor/baseline_models/temporal/pytorch/__init__.py
@@ -1,1 +1,0 @@
-"""Baseline pytorch models for temporal drift detection."""
